feat: allow case-insensitive `verify status` command, without being logged in
The MythX portal displays UUIDs in uppercase and therfore can’t be copy/pasted to the `verify status <uuid>` command.

This PR allows for case-insensitive entry of the UUID in to the `verify status <uuid>` command.

Additionally, running `verify status <uuid>` without first running a verify command would fail due to not being logged in.

This PR logs in when the `verify status <uuid>` command is run, so this command can be run *before* `verify`.